third_party/WebKit/LayoutTests/fullscreen/full-screen-iframe-zIndex-expected.txt

This tests that an element with a positive z-index appears behind the full screen iframe element. After entering full screen mode, the whole screen should be white. Click go full screen to run the test.

EVENT(webkitfullscreenchange)
END OF TEST
